fix(onchange): apply PR #198 review suggestions
- Clone `values` once, reuse for `ctx.value` (Gemini): avoids a redundant
  deep clone for the triggering field when values is large.
- Revoke warning sink on the success path too (Gemini): sink is now live
  only during the hook's synchronous span; background promises launched
  from inside `compute` can never push late warnings after the evaluator
  has moved on.
- Split `onchange-evaluator-errors.test.ts` further (CodeRabbit):
  549 > 500 — moved lookup/query error-surfacing block into new
  `onchange-evaluator-sanitation.test.ts`. Errors now 302 lines,
  sanitation 247, both under the cap. Shared `createSpyLogger` helper
  lifted into `onchange-evaluator-fixtures.ts` (no duplication).

No test behavior change; test count held at 4338 pass / 0 fail.
