+++ packages/core/__tests__/onchange-evaluator-errors.test.ts
@@ -1,53 +1,29 @@
 /**
- * Onchange evaluator — error surfacing, timeouts, warning sanitation, and
- * mutation safety.
+ * Onchange evaluator — baseline failures + Finding 3 timeout mutation safety.
  *
- * This file owns the regression tests for the CodeRabbit review of PR #191
- * (issue #192 / Spec 64 Phase 1):
+ * This file owns the baseline failure tests and the timeout-mutation
+ * regression tests from the CodeRabbit review of PR #191 (issue #192 /
+ * Spec 64 Phase 1):
  *
- *   - Finding 3: timed-out hooks must not mutate shared evaluation state.
- *   - Finding 4: raw internal error messages must not leak into the
- *     client-visible warnings array; they must go through `Logger.warn`.
- *   - Finding 5: `ctx.values` / `lookup` / `query` results are handed to
- *     hooks via a defensive clone so a misbehaving hook cannot mutate the
- *     shared accumulator seen by subsequent hooks.
+ *   - Finding 3: timed-out hooks must not mutate shared evaluation state
+ *     (late updates dropped, revocable sink suppresses late warnings).
+ *
+ * Related splits:
+ *   - onchange-evaluator-sanitation.test.ts — Finding 4 lookup/query error
+ *     sanitation and permission-check surfacing.
+ *   - onchange-evaluator-mutation.test.ts   — Finding 5 clone-boundary checks.
  */
 
 import { describe, expect, test } from "bun:test";
 import { createOnchangeEvaluator } from "../src/engine/onchange-evaluator";
 import type { EntityDefinition } from "../src/types/entity";
-import type { Logger } from "../src/types/logger";
 import {
   ACTOR,
   createFailingDataProvider,
+  createSpyLogger,
   createStubDataProvider,
   registerEntity,
 } from "./onchange-evaluator-fixtures";
-
-// ── Logger spy helper ──────────────────────────────────────
-
-interface LoggerCall {
-  level: "debug" | "info" | "warn" | "error";
-  message: string;
-  context?: Record<string, unknown>;
-}
-
-function createSpyLogger(): { logger: Logger; calls: LoggerCall[] } {
-  const calls: LoggerCall[] = [];
-  const make =
-    (level: LoggerCall["level"]) => (message: string, context?: Record<string, unknown>) => {
-      calls.push({ level, message, context });
-    };
-  return {
-    calls,
-    logger: {
-      debug: make("debug"),
-      info: make("info"),
-      warn: make("warn"),
-      error: make("error"),
-    },
-  };
-}
 
 // ── Baseline error handling ────────────────────────────────
 
@@ -321,229 +297,6 @@ describe("createOnchangeEvaluator — timeout mutation safety (Finding 3)", () =
   });
 });
 
-// ── Finding 4: lookup/query error sanitation ───────────────
-
-describe("createOnchangeEvaluator — lookup/query error surfacing", () => {
-  test("DataProvider lookup error: sanitized warning, raw detail goes to Logger (Finding 4)", async () => {
-    const entity: EntityDefinition = {
-      name: "line",
-      fields: {
-        product_id: { type: "string" },
-        unit_price: { type: "number" },
-        status: { type: "string" },
-      },
-      onchange: {
-        product_id: {
-          updates: ["unit_price", "status"],
-          compute: async (ctx) => {
-            const price = await ctx.lookup("product", ctx.value as string, "price");
-            return {
-              unit_price: price,
-              status: "ok",
-            };
-          },
-        },
-      },
-    };
-    const { logger, calls } = createSpyLogger();
-    const evaluator = createOnchangeEvaluator({
-      entityRegistry: registerEntity(entity),
-      dataProvider: createFailingDataProvider("SQL error: detail=xyz"),
-      logger,
-    });
-
-    const result = await evaluator.evaluate({
-      entityName: "line",
-      changedField: "product_id",
-      values: { product_id: "p1" },
-      actor: ACTOR,
-    });
-
-    expect(result.updates.status).toBe("ok");
-    expect(result.updates.unit_price).toBeUndefined();
-
-    // Finding 4 — user-facing warning does NOT contain any internal detail.
-    const lookupWarnings = result.warnings.filter((w) => w.includes('lookup on "product" failed'));
-    expect(lookupWarnings.length).toBe(1);
-    expect(lookupWarnings[0]).not.toContain("SQL error");
-    expect(lookupWarnings[0]).not.toContain("detail=xyz");
-
-    // …but the Logger received the real error with full context.
-    const loggedLookup = calls.find(
-      (c) => c.level === "warn" && c.message === "onchange: lookup failed",
-    );
-    expect(loggedLookup).toBeDefined();
-    expect(loggedLookup?.context?.error).toBe("SQL error: detail=xyz");
-    expect(loggedLookup?.context?.entity).toBe("product");
-    expect(loggedLookup?.context?.field).toBe("price");
-  });
-
-  test("DataProvider query error: sanitized warning, raw detail goes to Logger", async () => {
-    const entity: EntityDefinition = {
-      name: "line",
-      fields: {
-        product_id: { type: "string" },
-        unit_price: { type: "number" },
-      },
-      onchange: {
-        product_id: {
-          updates: ["unit_price"],
-          compute: async (ctx) => {
-            const list = await ctx.query("product", { kind: "widget" });
-            return { unit_price: list.length };
-          },
-        },
-      },
-    };
-    const { logger, calls } = createSpyLogger();
-    const evaluator = createOnchangeEvaluator({
-      entityRegistry: registerEntity(entity),
-      dataProvider: createFailingDataProvider("connection reset by peer"),
-      logger,
-    });
-
-    const result = await evaluator.evaluate({
-      entityName: "line",
-      changedField: "product_id",
-      values: { product_id: "p1" },
-      actor: ACTOR,
-    });
-    expect(result.updates.unit_price).toBe(0);
-
-    const queryWarnings = result.warnings.filter((w) => w.includes('query on "product" failed'));
-    expect(queryWarnings.length).toBe(1);
-    expect(queryWarnings[0]).not.toContain("connection reset");
-
-    const loggedQuery = calls.find(
-      (c) => c.level === "warn" && c.message === "onchange: query failed",
-    );
-    expect(loggedQuery).toBeDefined();
-    expect(loggedQuery?.context?.error).toBe("connection reset by peer");
-  });
-
-  test("permission-check failure: sanitized warning + Logger entry (Finding 4)", async () => {
-    const entity: EntityDefinition = {
-      name: "line",
-      fields: {
-        product_id: { type: "string" },
-        unit_price: { type: "number" },
-      },
-      onchange: {
-        product_id: {
-          updates: ["unit_price"],
-          compute: async (ctx) => ({
-            unit_price: await ctx.lookup("product", ctx.value as string, "price"),
-          }),
-        },
-      },
-    };
-    const { logger, calls } = createSpyLogger();
-    const evaluator = createOnchangeEvaluator({
-      entityRegistry: registerEntity(entity),
-      dataProvider: createStubDataProvider({
-        records: { product: { p1: { id: "p1", price: 42 } } },
-      }),
-      checkReadPermission: () => {
-        throw new Error("internal ACL lookup exploded: role=admin");
-      },
-      logger,
-    });
-
-    const result = await evaluator.evaluate({
-      entityName: "line",
-      changedField: "product_id",
-      values: { product_id: "p1" },
-      actor: ACTOR,
-    });
-
-    const warning = result.warnings.find((w) => w.includes("permission check failed"));
-    expect(warning).toBeDefined();
-    expect(warning).not.toContain("internal ACL lookup exploded");
-    expect(warning).not.toContain("role=admin");
-
-    const loggedPerm = calls.find(
-      (c) => c.level === "warn" && c.message === "onchange: read-permission check failed",
-    );
-    expect(loggedPerm).toBeDefined();
-    expect(loggedPerm?.context?.error).toContain("internal ACL lookup exploded");
-  });
-
-  test("permission denial stays as a permission-warning (not a lookup-warning)", async () => {
-    const entity: EntityDefinition = {
-      name: "line",
-      fields: {
-        product_id: { type: "string" },
-        unit_price: { type: "number" },
-      },
-      onchange: {
-        product_id: {
-          updates: ["unit_price"],
-          compute: async (ctx) => ({
-            unit_price: await ctx.lookup("product", ctx.value as string, "price"),
-          }),
-        },
-      },
-    };
-    const evaluator = createOnchangeEvaluator({
-      entityRegistry: registerEntity(entity),
-      dataProvider: createStubDataProvider({
-        records: { product: { p1: { id: "p1", price: 42 } } },
-      }),
-      checkReadPermission: () => false,
-    });
-
-    const result = await evaluator.evaluate({
-      entityName: "line",
-      changedField: "product_id",
-      values: { product_id: "p1" },
-      actor: ACTOR,
-    });
-    // Permission path is taken (not the try/catch fallback) — so we see the
-    // access-denied warning and NOT a lookup-failed warning.
-    expect(result.warnings.some((w) => w.includes('Access to "product" denied'))).toBe(true);
-    expect(result.warnings.some((w) => w.includes('lookup on "product" failed'))).toBe(false);
-  });
-
-  test("collapses repeated identical lookup errors across chained hooks", async () => {
-    const entity: EntityDefinition = {
-      name: "chain",
-      fields: {
-        a: { type: "string" },
-        b: { type: "string" },
-        c: { type: "string" },
-      },
-      onchange: {
-        a: {
-          updates: ["b"],
-          compute: async (ctx) => {
-            await ctx.lookup("product", "x", "price");
-            return { b: "next" };
-          },
-        },
-        b: {
-          updates: ["c"],
-          compute: async (ctx) => {
-            await ctx.lookup("product", "y", "price");
-            return { c: "done" };
-          },
-        },
-      },
-    };
-    const evaluator = createOnchangeEvaluator({
-      entityRegistry: registerEntity(entity),
-      dataProvider: createFailingDataProvider("DB connection reset"),
-    });
-
-    const result = await evaluator.evaluate({
-      entityName: "chain",
-      changedField: "a",
-      values: { a: "start" },
-      actor: ACTOR,
-    });
-    expect(result.updates).toEqual({ b: "next", c: "done" });
-    const lookupWarnings = result.warnings.filter((w) => w.includes('lookup on "product" failed'));
-    expect(lookupWarnings.length).toBe(1);
-  });
-});
-
+// Lookup/query/permission error-sanitation tests live in
+// onchange-evaluator-sanitation.test.ts.
 // Mutation-safety tests live in onchange-evaluator-mutation.test.ts.
