+++ packages/core/__tests__/onchange-evaluator-fixtures.ts
@@ -14,9 +14,39 @@ import type { DataProvider } from "../src/engine/action-engine";
 import { createEntityRegistry } from "../src/entity/entity-registry";
 import type { Actor } from "../src/types/action";
 import type { EntityDefinition } from "../src/types/entity";
+import type { Logger } from "../src/types/logger";
 
 /** Standard non-admin actor used by every test. */
 export const ACTOR: Actor = { type: "human", id: "u1", groups: ["user"] };
+
+/** Captured logger invocation used by sanitation-focused tests. */
+export interface LoggerCall {
+  level: "debug" | "info" | "warn" | "error";
+  message: string;
+  context?: Record<string, unknown>;
+}
+
+/**
+ * Spy logger that records every call. Used by the Finding 4 sanitation tests
+ * to assert that raw internal errors are routed to the server log rather than
+ * leaked into the client-visible warnings array.
+ */
+export function createSpyLogger(): { logger: Logger; calls: LoggerCall[] } {
+  const calls: LoggerCall[] = [];
+  const make =
+    (level: LoggerCall["level"]) => (message: string, context?: Record<string, unknown>) => {
+      calls.push({ level, message, context });
+    };
+  return {
+    calls,
+    logger: {
+      debug: make("debug"),
+      info: make("info"),
+      warn: make("warn"),
+      error: make("error"),
+    },
+  };
+}
 
 /** Minimal in-memory data provider for lookup/query tests. */
 export function createStubDataProvider(seed?: {
